Added incomplete Lexer object and sanitycheck method
Framework for lexer added, still need to make variables, chars, digits and other symbols work as well as write more test cases.
Made a 'sanityCheck' method to the compiler which reviews the input source code and ensures that it at least starts with a { and ends with a }$. It will correct these mistakes and warn the user about it. Doing this does offset the row/column feedback that the user sees which I'll need to fix. Will probably directly edit the text to fix but I'm going to see if I can find another solution that doesn't involve editing the sourcecode automatically because that isn't good UX for coders.