Added dynamic tokens for characters and digits
Added the ID and DIGIT token type that the lexer applies. These do not have a static lexeme, instead using a 'symbol' property for a-z or 0-9. This should allow me to create variables with multiple characters more easily if I wanted to later on with only a few small tweaks.