Invisible Char Output Aliasing
Tabs and newlines do not show up in the output log because there is nothing visible to see, this replaces the output with the word 'tab' or 'space'
Funfact: It is actually extremely difficult to put a tab or newline in the code menu atm because tabbing will just move to the next form. I realized this was a thing when I was trying to cut and paste the example test case code in the program 1 pdf.