Added Parser code structure
added the parser object, hooking it up to the compiler object and feeding it the token stream. Created the tree and node classes. Set up the error handling for the parser so that it will only return the first problem. Funnily enough it may be easier to have it return multiple errors but I can try that later if I want ig.